Removes all direct references between entities.
All of the entity types (Season, Conference, Division, Team, Game) now
only hold the ID of any associated entity, instead of holding direct
object references. For example, a Team no longer holds a list of the
games associated with that team. Instead, a Game has the IDs of the two
teams that are associated with it. This makes it much easier to
retrieving individual objects from the data store, without having to
retrieve the entire dataset.